# Conflicts:
#	uikit/src/main/res/drawable/text_input_background_error.xml
#	uikit/src/main/res/drawable/text_input_background_success.xml
#	uikit/src/main/res/values/styles.xml

+++ app/src/main/java/app/cashadvisor/common/utill/exceptions/LoginException.kt
@@ -44,7 +44,7 @@ sealed class LoginException(
         class UnauthorizedWrongConfirmationCode(
             override val message: String,
             val remainingAttempts: Int,
-            val lockDuration: Int,
+            val lockDuration: Long,
             val statusCode: Int
         ) : LoginCodeConfirmation(message)
